extract fetch call from login form into authenticate file
