+++ src/components/login/login_form.js
@@ -1,28 +1,7 @@
 import React, { Component } from 'react';
 import { View, TextInput, StyleSheet, TouchableOpacity, Text, AsyncStorage, Keyboard } from 'react-native';
 import styles from '../../../src/styles/login/loginFormStyles'
-
-const doSignIn = (email, password) => {
-  return fetch('https://vast-wildwood-58678.herokuapp.com/authenticate', {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify({
-      email: email,
-      password: password
-    })
-  })
-  .then((response) => response.json())
-  .then((responseJson) => {
-    return responseJson
-  })
-  .catch((error) => {
-    alert(error)
-  })
-}
-
+import doSignIn from '../../../src/fetch/authenticate'
 
 export default class LoginForm extends Component {
   constructor(props){
